requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.8.1
attrs==24.2.0
automium==0.2.6
automium_web==0.1.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
crontab==1.0.1
Django==5.1.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Flask==3.0.3
frozenlist==1.5.0
geoip2==4.8.0
h11==0.14.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
legacy==0.1.7
lxml==5.3.0
macholib==1.16.3
MarkupSafe==2.1.5
maxminddb==2.6.2
monitoring==0.9.12
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
propcache==0.2.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.9.0
PySocks==1.7.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
redis-cli==1.0.1
requests==2.32.3
selenium==4.25.0
setuptools==75.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
views==0.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.17.1

self~=2020.12.3